remove fpu/softfloat.h, reduce include deprendencies, fix warnings regenerate dep lists in all Makefile.in

+++ bochs/cpu/avx/avx512_rsqrt14.cc
@@ -8234,6 +8234,7 @@ static const Bit16u rsqrt14_table1[32768] = {
     0x6a0e, 0x6a0d, 0x6a0c, 0x6a0c, 0x6a0b, 0x6a0a, 0x6a09, 0x6a09  // 32760
 };
 
+#include "softfloat3e/include/softfloat.h"
 #include "fpu/softfloat-specialize.h"
 #include "simd_int.h"
 
@@ -8427,7 +8428,7 @@ void BX_CPP_AttrRegparmN(1) BX_CPU_C::VRSQRT14PS_MASK_VpsWpsR(bxInstruction_c *i
 
   for (unsigned n=0, tmp_mask = mask; n < DWORD_ELEMENTS(len); n++, tmp_mask >>= 1) {
     if (tmp_mask & 0x1)
-      op.vmm32u(n) = approximate_rsqrt14(op.vmm32u(n), MXCSR.get_DAZ());
+      op.vmm32u(n) = approximate_rsqrt14((float32) op.vmm32u(n), MXCSR.get_DAZ());
     else
       op.vmm32u(n) = 0;
   }
@@ -8452,7 +8453,7 @@ void BX_CPP_AttrRegparmN(1) BX_CPU_C::VRSQRT14PD_MASK_VpdWpdR(bxInstruction_c *i
 
   for (unsigned n=0, tmp_mask = mask; n < QWORD_ELEMENTS(len); n++, tmp_mask >>= 1) {
     if (tmp_mask & 0x1)
-      op.vmm64u(n) = approximate_rsqrt14(op.vmm64u(n), MXCSR.get_DAZ());
+      op.vmm64u(n) = approximate_rsqrt14((float64) op.vmm64u(n), MXCSR.get_DAZ());
     else
       op.vmm64u(n) = 0;
   }
@@ -8513,7 +8514,6 @@ void BX_CPP_AttrRegparmN(1) BX_CPU_C::VRSQRTSH_MASK_VshHphWshR(bxInstruction_c *
 
   if (! i->opmask() || BX_SCALAR_ELEMENT_MASK(i->opmask())) {
     float16 op2 = BX_READ_XMM_REG_LO_WORD(i->src2());
-
     op1.xmm16u(0) = approximate_rsqrt14(op2, MXCSR.get_DAZ());
   }
   else {
@@ -8535,7 +8535,7 @@ void BX_CPP_AttrRegparmN(1) BX_CPU_C::VRSQRTPH_MASK_VphWphR(bxInstruction_c *i)
 
   for (unsigned n=0, tmp_mask = mask; n < WORD_ELEMENTS(len); n++, tmp_mask >>= 1) {
     if (tmp_mask & 0x1)
-      op.vmm16u(n) = approximate_rsqrt14(op.vmm16u(n), MXCSR.get_DAZ());
+      op.vmm16u(n) = approximate_rsqrt14((float16) op.vmm16u(n), MXCSR.get_DAZ());
     else
       op.vmm16u(n) = 0;
   }
